gnl ok after tester
fixed leaks

+++ get_next_line.h
@@ -24,7 +24,7 @@
 char	*get_next_line(int fd);
 char	*get_line(char **str);
 int		has_line(char *str);
-size_t	ft_strlen(char *str);
+int		ft_strlen(char *str);
 char	*ft_strjoin(char *str1, char *str2);
 char	*ft_strndup(char *str, int start, int end);
